Introduce BlockchainApi#getBoxById method
